requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
sqlalchemy==2.0.15
psycopg2-binary==2.9.7
nltk==3.8.1
pydub==0.25.1
speechrecognition==3.10.0
pydantic==1.10.7
pyttsx3==2.90
beautifulsoup4==4.12.2
requests==2.28.2
python-dotenv==1.0.0
pydub==0.25.1
ffmpeg-python==0.2.0